Deprecate the `java_tests` alias in favor of `junit_tests`.
Testing Done:
CI passed: https://travis-ci.org/pantsbuild/pants/builds/168784588

Reviewed at https://rbcommons.com/s/twitter/r/4322/
