+++ src/python/pants/backend/jvm/targets/scala_library.py
@@ -7,7 +7,7 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 
 from pants.backend.jvm.subsystems.scala_platform import ScalaPlatform
 from pants.backend.jvm.targets.exportable_jvm_library import ExportableJvmLibrary
-from pants.backend.jvm.targets.java_tests import JavaTests
+from pants.backend.jvm.targets.junit_tests import JUnitTests
 from pants.base.exceptions import TargetDefinitionException
 from pants.build_graph.address import Address
 
@@ -25,7 +25,7 @@ class ScalaLibrary(ExportableJvmLibrary):
   """
 
   default_sources_globs = '*.scala'
-  default_sources_exclude_globs = JavaTests.scala_test_globs
+  default_sources_exclude_globs = JUnitTests.scala_test_globs
 
   @classmethod
   def subsystems(cls):
