+++ testprojects/src/java/org/pantsbuild/testproject/junit/earlyexit/BUILD
@@ -1,4 +1,4 @@
-java_tests(name='tests',
+junit_tests(name='tests',
   sources=['AllTests.java'],
   dependencies=['3rdparty:junit'],
 )
