+++ testprojects/src/java/org/pantsbuild/testproject/junit/mixed/tests/org/pantsbuild/tmp/tests/BUILD
@@ -1,4 +1,4 @@
-java_tests(
+junit_tests(
   sources=[
     'AllTestsBase.java',
     'AllTests.java',
