+++ tests/java/org/pantsbuild/tools/junit/impl/BUILD
@@ -1,7 +1,7 @@
 # Copyright 2015 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-java_tests(
+junit_tests(
   name='junit',
   dependencies=[
     '3rdparty:guava',
